feat[Shares Calculation]: Created the Endpoint and the function to calculate the current owned shares and values by employeeId

+++ migrations/20220702161902-create-contract.js
@@ -13,7 +13,7 @@ module.exports = {
         allowNull: false
       },
       companyValuation: {
-        type: Sequelize.NUMERIC(12, 2),
+        type: Sequelize.FLOAT,
         allowNull: false
       },
       totalCompanyShares: {
@@ -21,7 +21,7 @@ module.exports = {
         allowNull: false
       },
       grantedShares: {
-        type: Sequelize.NUMERIC(12, 2),
+        type: Sequelize.FLOAT,
         allowNull: false
       },
       cliffDate: {
